+++ DeviantartApi/Objects/SubmitResult.cs
@@ -5,12 +5,12 @@ namespace DeviantartApi.Objects
     public class SubmitResult : BaseObject
     {
         [JsonProperty("itemid")]
-        public int ItemId { get; set; }
+        public long ItemId { get; set; }
 
         [JsonProperty("stack")]
         public string Stack { get; set; }
 
         [JsonProperty("stackid")]
-        public int StackId { get; set; }
+        public long StackId { get; set; }
     }
 }
